requirements.txt
Pillow==8.3.2
selenium==4.1.5